+++ guillotina/contrib/catalog/pg.py
@@ -87,6 +87,7 @@ WHERE
 
 
 class ParsedQueryInfo(BasicParsedQueryInfo):
+    sort_on_fields: bool
     wheres: typing.List[typing.Any]
     wheres_arguments: typing.List[typing.Any]
     selects: typing.List[str]
@@ -140,6 +141,7 @@ class Parser(BaseParser):
             return self.process_compound_field(field, value, " AND ")
 
         result: typing.Any = value
+
         operator = "="
         if field.endswith("__not"):
             operator = "!="
@@ -202,7 +204,7 @@ class Parser(BaseParser):
             operator = "?|"
 
         pg_index = get_pg_index(field)
-        return pg_index.where(result, operator), [result], pg_index.select()
+        return pg_index.where(result, operator), [result], pg_index.select(), field
 
     def __call__(self, params: typing.Dict) -> ParsedQueryInfo:
         query_info = super().__call__(params)
@@ -210,21 +212,27 @@ class Parser(BaseParser):
         arguments = []
         selects = []
         selects_arguments = []
+        sort_field = query_info.get("sort_on", None)
+        sort_on_fields = False
         for field, value in query_info["params"].items():
             result = self.process_queried_field(field, value)
             if result is None:
                 continue
-            sql, values, select = result
+            sql, values, select, field = result
+            if sort_field is not None and field == sort_field:
+                sort_on_fields = True
             wheres.append(sql)
             arguments.extend(values)
             if select:
                 selects.extend(select)
                 selects_arguments.extend(values)
 
+
         return typing.cast(
             ParsedQueryInfo,
             dict(
                 query_info,
+                sort_on_fields=sort_on_fields,
                 wheres=wheres,
                 wheres_arguments=arguments,
                 selects=selects,
@@ -365,7 +373,7 @@ using gin(to_tsvector('english', json->>'{sqlq(self.name)}'));"""
         return f"""
 to_tsvector('english', json->>'{sqlq(self.name)}') @@ plainto_tsquery(${{arg}}::text)"""
 
-    def order_by(self, direction="ASC"):
+    def order_by_score(self, direction="ASC"):
         return f"order by {sqlq(self.name)}_score {sqlq(direction)}"
 
     def select(self):
@@ -495,6 +503,7 @@ class PGSearchUtility(DefaultSearchUtility):
         sql_arguments = []
         sql_wheres = []
         arg_index = 1
+
         for idx, select in enumerate(query["selects"]):
             select_fields.append(select.format(arg=arg_index))
             sql_arguments.append(query["selects_arguments"][idx])
@@ -520,6 +529,7 @@ class PGSearchUtility(DefaultSearchUtility):
             raise TransactionNotFound()
         sql_wheres.extend(self.get_default_where_clauses(container))
 
+        order = order_by_index.order_by_score(query["sort_dir"]) if query["sort_on_fields"] else order_by_index.order_by(query["sort_dir"])
         sql = """select {} {}
                  from {}
                  where {}
@@ -529,7 +539,7 @@ class PGSearchUtility(DefaultSearchUtility):
             ",".join(select_fields),
             sqlq(txn.storage.objects_table_name),
             " AND ".join(sql_wheres),
-            "" if distinct else order_by_index.order_by(query["sort_dir"]),
+            "" if distinct else order,
             sqlq(query["size"]),
             sqlq(query["_from"]),
         )