+++ guillotina/tests/test_catalog.py
@@ -334,6 +334,39 @@ async def test_fulltext_query_pg_catalog(container_requester):
 
 @pytest.mark.app_settings(PG_CATALOG_SETTINGS)
 @pytest.mark.skipif(NOT_POSTGRES, reason="Only PG")
+async def test_fulltext_query_pg_catalog_order(container_requester):
+    from guillotina.contrib.catalog.pg import PGSearchUtility
+
+    async with container_requester as requester:
+        await requester(
+            "POST",
+            "/db/guillotina/",
+            data=json.dumps({"@type": "Item", "id": "item1", "title": "Something interesting"}),
+        )
+        await requester(
+            "POST",
+            "/db/guillotina/",
+            data=json.dumps({"@type": "Item", "title": "Something else something foobar something", "id": "item2"}),
+        )
+
+        async with requester.db.get_transaction_manager() as tm, await tm.begin():
+            test_utils.login()
+            root = await tm.get_root()
+            container = await root.async_get("guillotina")
+
+            util = PGSearchUtility()
+            await util.initialize()
+            results = await util.search(container, {"title": "something", "_sort_asc": "title"})
+            assert results['items'][0]['title'] == 'Something interesting'
+            assert len(results["items"]) == 2
+
+            results = await util.search(container, {"_sort_asc": "title"})
+            assert results['items'][0]['title'] == 'Something else something foobar something'
+            assert len(results["items"]) == 2
+
+
+@pytest.mark.app_settings(PG_CATALOG_SETTINGS)
+@pytest.mark.skipif(NOT_POSTGRES, reason="Only PG")
 async def test_build_pg_query(dummy_guillotina):
     from guillotina.contrib.catalog.pg import PGSearchUtility